Created Presenter, Model, View(interface), MainActivity Addition created CounterButton sealed class for identification clicked button and work with ligic of application.
