+++ app/src/main/AndroidManifest.xml
@@ -4,6 +4,7 @@
     package="ru.marslab.popularlibraries">
 
     <application
+        android:name=".App"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
