+++ app/src/main/java/ru/marslab/popularlibraries/App.kt
@@ -1,0 +1,13 @@
+package ru.marslab.popularlibraries
+
+import android.app.Application
+import ru.marslab.popularlibraries.presenter.MainPresenter
+import ru.marslab.popularlibraries.presenter.MainPresenterImpl
+
+class App : Application() {
+
+    private val presenter: MainPresenter = MainPresenterImpl()
+
+    fun getPresenter(): MainPresenter =
+        presenter
+}
